pyproject.toml
[project]
name = "second-brain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.0",
    "langchain>=1.0.3",
    "langchain-text-splitters>=1.0.0",
    "logfire>=1.0.0",
    "openai>=2.6.1",
    "opentelemetry-api>=1.24.0",
    # Note: opentelemetry-sdk and opentelemetry-exporter-otlp-proto-http are only needed
    # as fallback when logfire is disabled. Logfire already includes these.
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp-proto-http>=1.24.0",
    "pydantic>=2.12.3",
    "pydantic-ai>=1.9.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "guardrails-ai>=0.1.8",
]

[project.scripts]
second-brain = "second_brain.main:main"

[build-system]
requires = ["uv_build>=0.9.3,<0.10.0"]
build-backend = "uv_build"
